Changed semantics of Process.Terminate: The process is not immediately marked as exited;  It is marked as exited when the callback is received. Added tests for process termination. Tracking of COM objects is no longer outputted to the console.
git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/trunk@3132 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61

+++ src/AddIns/Misc/Debugger/Debugger.Tests/Project/Src/DebuggerTestsBase.cs
@@ -78,7 +78,7 @@ namespace Debugger.Tests
 			CheckXmlOutput();
 		}
 		
-		void CheckXmlOutput()
+		protected void CheckXmlOutput()
 		{
 			string startMark = "#if EXPECTED_OUTPUT\r\n";
 			string endMark = "#endif // EXPECTED_OUTPUT";
@@ -307,11 +307,15 @@ namespace Debugger.Tests
 			string path = Path.GetTempPath();
 			path = Path.Combine(path, "SharpDevelop");
 			path = Path.Combine(path, "DebuggerTests");
-			path = Path.Combine(path, md5);
+			path = Path.Combine(path, testName + "." + md5);
 			Directory.CreateDirectory(path);
 			
 			string codeFilename = Path.Combine(path, testName);
 			string exeFilename = Path.Combine(path, testName.Replace(".cs", ".exe"));
+			
+			if (File.Exists(exeFilename)) {
+				return exeFilename;
+			}
 			
 			StreamWriter file = new StreamWriter(codeFilename);
 			file.Write(code);
